# src/main/ressource/openapi.yaml
openapi: 3.1.0
info:
  title: Agricultural federation API
  description: |
    # Versioning
    This API uses semantic versioning (MAJOR.MINOR.PATCH). Available versions are : 
    - v0.0.1 (2026-04-21) : POST /collectivities for collectivities creation and POST /members for members creation.
    - v0.0.2 (2026-04-22) : PUT /collectivities/{id}/informations for collectivity information update.
    - v0.0.3 (2026-04-23) : New features about membership fees, collectivities transactions and its financial accounts, and members payments.
    - v0.0.4 (2026-04-23) : GET /collectivities/{id} and  GET /collectivities/{id}/financialAccounts
    - v0.0.5 (2026-05-05) : Partial features collectivity statistics.
    - v0.0.6 (2026-05-05) : New features about collectivities activities and assiduity.
  version: 0.0.6
servers:
  - url: 'https://localhost:8080'
paths:
  '/collectivities':
    post:
      summary: Create list of collectivities
      description: |
        During collectivity request body creation, only members ID are provided to identify members. 
        But the response content includes members information, not only their IDs.
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateCollectivity'
      responses:
        201:
          description: List of created collectivities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collectivity'
        400:
          description: Collectivity without federation approval or structure missing.
        404:
          description: Member not found.
  '/collectivities/{id}':
    get:
      summary: Retrieve collectivity by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        200:
          description: Collectivity retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collectivity'
        404:
          description: Collectivity not found
...
  '/collectivities/{id}/informations':
    put:
      summary: Attribute collectivity informations to specific collectivity
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectivityInformation'
      responses:
        200:
          description: Collectivity with updated information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collectivity'
        400:
          description: Either given number or name is already used by other collectivity.
        404:
          description: Collectivity not found.
...
components:
  schemas:
...
    CreateCollectivity:
      type: object
      properties:
        location:
          type: string
        members:
          type: array
          items:
            schema:
              $ref: '#/components/schemas/MemberIdentifier'
        federationApproval:
          type: boolean
        structure:
          $ref: '#/components/schemas/CreateCollectivityStructure'
    CreateCollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/MemberIdentifier'
        vicePresident:
          $ref: '#/components/schemas/MemberIdentifier'
        treasurer:
          $ref: '#/components/schemas/MemberIdentifier'
        secretary:
          $ref: '#/components/schemas/MemberIdentifier'
    CollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/Member'
        vicePresident:
          $ref: '#/components/schemas/Member'
        treasurer:
          $ref: '#/components/schemas/Member'
        secretary:
          $ref: '#/components/schemas/Member'
    CollectivityInformation:
      type: object
      properties:
        name:
          type: string
        number:
          type: int
    Collectivity:
      type: object
      allOf:
        - $ref: '#/components/schemas/CollectivityInformation'
      properties:
        id:
          type: string
        location:
          type: string
        structure:
          $ref: '#/components/schemas/CollectivityStructure'
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
...
    MemberIdentifier:
      type: string
...
    MemberInformation:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
        profession:
          type: string
        phoneNumber:
          type: int
        email:
          type: string
        occupation:
          $ref: '#/components/schemas/MemberOccupation'
    Member:
      type: object
      allOf:
        - $ref: '#/components/schemas/MemberInformation'
      properties:
        id:
          $ref: '#/components/schemas/MemberIdentifier'
        referees:
          type: array
          items:
            $ref: '#/components/schemas/Member'
...
    Gender:
      type: string
      enum:
        - MALE
        - FEMALE
    MemberOccupation:
      type: string
      enum:
        - JUNIOR
        - SENIOR
        - SECRETARY
        - TREASURER
        - VICE_PRESIDENT
        - PRESIDENT
...